+++ src/solver/main/adequacy.cpp
@@ -25,7 +25,7 @@
 ** SPDX-License-Identifier: licenceRef-GPL3_WITH_RTE-Exceptions
 */
 
-#include "../application.h"
+#include "../application/application.h"
 #include "../simulation/solver.h"
 #include "../simulation/adequacy.h"
 #include <antares/benchmarking.h>
